+++ terraform/node_resource_apply.go
@@ -108,6 +108,6 @@ func (n *NodeApplyableResource) Execute(ctx EvalContext, op walkOperation) error
 		return nil
 	}
 
-	err := n.WriteResourceState(ctx, n.Addr)
+	err := n.writeResourceState(ctx, n.Addr)
 	return err
 }
